+++ BusinessObjects/Dto/BuyBack/BuybackCustomerDto.cs
@@ -1,0 +1,8 @@
+namespace BusinessObjects.Dto.BuyBack;
+
+public class BuybackCustomerDto
+{
+    public string FullName { get; set; }
+    public string Phone { get; set; }
+    public string Address { get; set; }
+}
