+++ BusinessObjects/Dto/BuyBack/JewelryMaterialDto.cs
@@ -1,0 +1,9 @@
+namespace BusinessObjects.Dto.BuyBack;
+
+public class JewelryMaterialDto
+{
+    public string GoldId { get; set; }
+    public string StoneId { get; set; }
+    public int GoldQuantity { get; set; }
+    public int StoneQuantity { get; set; }
+}
