+++ BusinessObjects/Dto/Dashboard/RevenueByProductTypeDto.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BusinessObjects.Dto.Revenue
+﻿namespace BusinessObjects.Dto.Dashboard
 {
     public class RevenueByProductTypeDto
     {
