+++ Services/Interface/IDashboardService.cs
@@ -1,5 +1,4 @@
 ﻿using BusinessObjects.Dto.Dashboard;
-using BusinessObjects.Dto.Revenue;
 using BusinessObjects.Models;
 using System;
 using System.Collections.Generic;
